StatModifier value types fixed and System.Random

+++ Scripts/StatModifier/Common/IStatModifierData.cs
@@ -2,6 +2,6 @@ public interface IStatModifierData
 {
     StatType StatType { get; }
     StatModifierType StatModifierType { get; }
-    int MinValue { get; }
-    int MaxValue { get; }
+    float MinValue { get; }
+    float MaxValue { get; }
 }
